integrations/telegram-bots/requirements.txt
python-telegram-bot>=21.0
aiohttp>=3.9
structlog>=24.0
